refactor(orchestrator): extract the Verify-exit release gate to release_transition.py; make the parallel suite load-tolerant
- _enforce_app_release_transition_inner (~420 lines) moves out of core.py
  (9822 -> 9680) as a module-level function that takes the orchestrator
  explicitly; core keeps a one-call forward and the validate_release_contract /
  resolve_target_release_identity test seams via _core().
- pytest -n auto: the worktree-script subprocess timeout is 60 s (was 5),
  the release-reconcile worker-start wait is 10 s (was 1), the concurrent
  reservation barrier/join are 30 s/60 s (were 5/10).
- board_health: only the Document lane crosses into a thread to aggregate
  stats.jsonl, so dispatch timing for every other lane is unchanged (the
  eager-task dispatch test caught the regression).
- HANDOFF.md rewritten for the 2026-09-12 pass; architecture table gains
  release_transition.py.
