+++ docs/architecture.md
@@ -114,6 +114,7 @@ modules.
 | `contracts.py` | stage-contract evaluation for pipeline headings and evidence gates |
 | `run_registry.py` | SQLite WAL run leases, dead-owner reclaim, persisted retry / pause / budget flags |
 | `attempt.py` | the worker body (`run_agent_attempt`): startup, backend start, session open, turn loop, cleanup, exit hand-off; phases pass state explicitly (`_AttemptState`, `_AttemptExit`, `_TurnFlow`) |
+| `release_transition.py` | the Verify-exit release gate for `app-release` work (`enforce_app_release_transition_inner`): gate bind/recover, contract validation, GREEN approval or repairable-RED hand-off; takes the orchestrator explicitly |
 | `worker_exit.py` | the worker-exit state machine (`handle_worker_exit`): exit accounting, Done post-processing, continuation / retry / escalation, `worker_exit` emit; takes the orchestrator explicitly |
 | `core.py` | the `Orchestrator` class, tick loop, dispatch, health, pause/resume, Document skip |
 | `__init__.py` | re-exports + monkeypatch-target bindings |
