+++ src/symphony/orchestrator/attempt.py
@@ -451,9 +451,11 @@ async def _open_session(orch: Orchestrator, st: _AttemptState) -> None:
     skill_context = await asyncio.to_thread(
         render_skill_block, st.cfg.workflow_path.parent, st.issue.skills
     )
-    board_health = await asyncio.to_thread(
-        orch._board_health_for_prompt, st.cfg, st.issue.state
-    )
+    board_health = ""
+    if orch._board_health_wanted(st.issue.state):
+        board_health = await asyncio.to_thread(
+            orch._board_health_for_prompt, st.cfg, st.issue.state
+        )
     first_prompt, _ = build_first_turn_prompt(
         prompt_template=st.cfg.prompt_template_for_state(st.issue.state),
         issue=st.issue,
